+++ src/components/NavBar.vue
@@ -1,5 +1,5 @@
 <template>
-  <nav class="navbar navbar-expand-md navbar-light bg-light">
+  <nav class="navbar navbar-light bg-light" :class="{ 'navbar-expand-md': !showHamburger }">
 
     <a v-if="molgenisMenu.navBarLogo" class="navbar-brand"
        :href="`/menu/main/${href(molgenisMenu.menu.items[0])}`">
@@ -14,7 +14,7 @@
     </button>
 
     <div class="collapse navbar-collapse" id="navbar-content">
-      <ul class="navbar-nav mr-auto">
+      <ul class="navbar-nav mr-auto" ref="mgNavBarNav">
 
         <template v-for="item in molgenisMenu.menu.items">
           <li :key="item.id" v-if="item.type === 'plugin' && item.id !== 'home'"
@@ -37,14 +37,15 @@
       </ul>
 
       <ul class="navbar-nav justify-content">
-        <li v-if="molgenisMenu.authenticated && languages.length > 1" class="nav-item">
+        <li v-if="molgenisMenu.authenticated && languages.length > 1 && selectedLanguage"
+            class="nav-item">
           <form id="language-form" class="navbar-form">
-            <select class="nav-link" v-model="selectedLanguage" @change="handleLanguageSelect">
+            <select class="nav-link" v-model="selectedLanguage.id" @change="handleLanguageSelect">
               <option
                 v-for="language in languages"
                 :key="language.id"
                 :value="language.id"
-                :selected="language.id === selectedLanguage"
+                :selected="language.id === selectedLanguage.id"
               >
                 {{ language.label }}
               </option>
@@ -81,7 +82,9 @@ import Vue from 'vue'
 import { MolgenisMenu } from '../types'
 import { href } from '../href'
 import DropDownItems from './DropDownItems'
-import api from '@molgenis/molgenis-api-client'
+import languageRepository from '../repository/LanguageRepository'
+import languageService from '../service/LanguageService'
+import eventUtilService from '../service/EventUtilService'
 
 export default Vue.extend({
   name: 'NavBar',
@@ -99,7 +102,10 @@ export default Vue.extend({
         display: 'inline-block',
         height: '100%',
         verticalAlign: 'middle'
-      } : undefined
+      } : undefined,
+      expectedNavHeight: null,
+      showHamburger: false,
+      dynamicHamburgerBreakpoint: null
     }
   },
   computed: {
@@ -127,26 +133,55 @@ export default Vue.extend({
       document.getElementById('logout-form').submit()
     },
     handleLanguageSelect () {
-      api.post('/plugin/useraccount/language/update?languageCode=' + this.selectedLanguage).then(
-        () => {
-          location.reload(true)
-        })
+      languageRepository.setSelectedLanguage(this.selectedLanguage.id).then(() => {
+        location.reload(true)
+      })
+    },
+    getClientWidth () {
+      return window.innerWidth || document.documentElement.clientWidth ||
+          document.body.clientWidth
+    },
+    handleResize () {
+      if (this.showHamburger) {
+        if (this.getClientWidth() > this.dynamicHamburgerBreakpoint) {
+          this.showHamburger = false
+        }
+      } else {
+        const actualNavHeight = this.$refs.mgNavBarNav.clientHeight
+        if (actualNavHeight > this.expectedNavHeight) {
+          this.dynamicHamburgerBreakpoint = this.getClientWidth()
+          this.showHamburger = true
+        }
+      }
+    },
+    debounce: eventUtilService.debounce,
+    getPixelValue (sourceObject, propertyName) {
+      return parseInt(sourceObject.getPropertyValue(propertyName), 10)
     }
   },
   mounted () {
+    const links = this.$refs.mgNavBarNav.getElementsByClassName('nav-link')
+    const linkStyleObject = window.getComputedStyle(links[0])
+    const lineHeight = this.getPixelValue(linkStyleObject, 'line-height')
+    const paddingTop = this.getPixelValue(linkStyleObject, 'padding-top')
+    const paddingBottom = this.getPixelValue(linkStyleObject, 'padding-bottom')
+    this.expectedNavHeight = lineHeight + paddingTop + paddingBottom
+
+    window.addEventListener('resize', this.debounce(this.handleResize, 100))
+
     if (this.molgenisMenu.authenticated) {
-      api.get('/api/v2/sys_Language?q=active==true').then(response => {
-        this.languages = response.items.map(item => {
-          const language = { id: item.code, label: item.name }
-          if (item.code === response.meta.languageCode) {
-            this.selectedLanguage = language.id
-          }
-          return language
-        })
-      }, error => {
-        console.error(error)
+      Promise.all([languageRepository.getActivelangueges(),
+        languageRepository.getSelectedlanguegeCode()]).then((results) => {
+        this.languages = results[0]
+        this.selectedLanguage = languageService.selectedLanguage(this.languages, results[1])
       })
     }
+  },
+  updated () {
+    this.handleResize()
+  },
+  beforeDestroy () {
+    window.removeEventListener('resize', this.handleResize)
   }
 })
 </script>