+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Dynamic (via js) menu collapse breakpoints (hamburger menu).
 ### Removed
 ### Changed
 ### Fixed